ms1/requirements.txt
kafka-python
faker
python-dotenv
requests

pandas
sqlalchemy
psycopg2-binary

fastapi
uvicorn